reorganizing examples, archives, folder structures
