+++ _archive/app.js
@@ -1,10 +1,7 @@
 /**
  * 180 Faces - Hardware Interface
  * Created by jefferson.wu on 4/11/16.
- * Namespaced + Jade
- *
- * TODO: create a new random object with local RGBholder values on 'connect', Destroy on 'disconnect'
- * TODO: environmental variables.  Use them.
+ * TODO: experimenting with socket.io namespaces
  */
 
 // SOCKET.IO BOILERPLATE
@@ -17,24 +14,6 @@ var io = require('socket.io')(server);
 
 var jade = require('jade');
 var colors = require('colors');
-var rgbHolder = {red: 0, green: 0, blue: 0};
-
-//serialport related
-var serialport = require('serialport');
-var SerialPort = serialport.SerialPort;
-var portName = process.argv[3];
-var portConfig = {
-    baudRate: 9600,
-    //calls myPort.on('data') when a newline is received
-    parser: serialport.parsers.readline('\n')
-};
-
-//open the serial port
-var myPort = new SerialPort(portName, portConfig);
-
-myPort.on('open', function(){
-    console.log('serial port opened, please wait 3 seconds');
-});
 
 // MIDDLEWARE
 app.use(express.static(__dirname + '/public'));
@@ -55,24 +34,7 @@ app.get('/viewer', function(request, response){
     response.sendFile(__dirname + '/public/debug-viewer.html');
 });
 
-/*SOCKET.IO NAMESPACE*/
-var namespaceString = 'servo';
-var nsp = io.of('/' + namespaceString);
-
-nsp.on('connection', function(socket){
-    console.log(socket.client.id.toString().blue + ' has connected to namespace: ' + namespaceString);
-
-    //work that servo (has to be under the namespace)
-    socket.on('servoData', function(data){
-        console.log('Message received: ' + data);
-        myPort.write(data);
-
-    });
-
-});
-/*SOCKET.IO NAMESPACE - END */
-
-/*SOCKET.IO*/
+// SOCKET.IO
 io.on('connection', function(socket){
     allClients.push(socket);
     console.log(socket.client.id.toString().blue + ' connected');
@@ -89,23 +51,6 @@ io.on('connection', function(socket){
         var stringObject = data;
         var dataObject = JSON.parse(stringObject);
         console.log(socket.client.id.toString().green + ': ' + dataObject.head + ', ' + dataObject.body);
-
-        //set the color
-        if(dataObject.head.toString() == 'slider1'){
-            rgbHolder.red = dataObject.body;
-        }
-        else if(dataObject.head.toString() == 'slider2'){
-            rgbHolder.green = dataObject.body;
-        }
-        else if(dataObject.head.toString() == 'slider3'){
-            rgbHolder.blue = dataObject.body;
-        }
-
-        console.log(rgbHolder);
-        //console.log(socket.client.id.toString().green + ': ' + dataObject.head + ', ' + dataObject.body);
-
-        //send to view
-        socket.broadcast.emit('viewUpdate', JSON.stringify(rgbHolder));
     });
 
     //on disconnect
@@ -114,13 +59,6 @@ io.on('connection', function(socket){
         allClients.splice(allClients.indexOf(socket.client.id), 1);
         //console.log(allClients);
         numberOfClients(allClients);
-
-        //reset the color on exit
-        rgbHolder.red = 0;
-        rgbHolder.green = 0;
-        rgbHolder.blue = 0;
-
-        console.log(rgbHolder);
     });
 
     // ==== SOCKET EVENTS - END ====
@@ -137,11 +75,6 @@ function initServer(port){
     console.log('Starting server on port ' + serverPort.rainbow);
 }
 
-function initSerialPort(){
-    console.log('port open');
-    console.log('baud rate: ' + myPort.options.baudRate);
-}
-
 /**
  * Return and log to console the number of connected clients.
  * @param allClientsArr The 'allClients' array
@@ -152,5 +85,6 @@ function numberOfClients(allClientsArr){
 }
 
 function updateViewer(dataObject){
+
 }
