+++ _examples/checksum/checksum.js
