+++ _examples/checksum/checksum2.js
