+++ _examples/serial-port/index.js
