+++ arduino/_archive/single_serial_signal/single_serial_signal.ino
