+++ arduino/faces_servo_mover/faces_servo_mover.ino
@@ -3,6 +3,7 @@
   ================================================
   Receive and react to a serial string.
   (for 180 faces)
+  - MERGED FROM node_tut.ino. 2016.APR.20
 
   Note:
   - nano resets when disconnected, micro keeps the connection alive.
@@ -17,103 +18,52 @@
 /*vars*/
 // ==== LED ====
 int ledPin = 13;
-// ==== Serial ====
-char val; //holds data from serial port
-int intVal; //holds integer from serial port
 
 // ==== Servo ====
 Servo myservo;
 int servoPos = 0; //keep track of the servo position for smoothing
 const int servoPin = 10;
-const char servoKey1 = 'a'; //zero position
-const char servoKey2 = 's'; //center position
-const char servoKey3 = 'd'; //max position
-const char servoKey4 = 'f'; //arbitrary position
-
-const char servoKey5 = 'g'; //test servos for-loop
-const char servoKey6 = 'z'; //smooth to close from any position
-const char servoKey7 = 'x'; //smooth to open from any position
-
-
 
 void setup() {
   pinMode(ledPin, OUTPUT);  //set LED to output
   Serial.begin(9600);       //pour a bowl of serial
+  Serial.setTimeout(10);   
   
   myservo.attach(servoPin); //attach servo to pin
   myservo.write(90);         //zero servo on startup
 
-  Serial.print("initialized! Input a value from 0-180. \r\n");
+  Serial.print("initialized! \r\n");
 }
 
 void loop() {
-  serialInput();
-  //serialEvaluation();
+  evaluateSerial();
   delay(100);
 }
 
 
-
 /*============= FUNCTIONS ============*/
-
-void serialInput() {
-  while(Serial.available() > 0) {
-    
-    int value = Serial.parseInt();
-
-    //end of line...
-    if(Serial.read() == '\n'){
-
-      
-
-      //move servo
-      if(value < 0) {
-        Serial.print("value too low! \r\n");
-      }
-      else if(value > 180){
-        Serial.print("value too high! \r\n");
-      }
-      else {
-        //write back to console
-        Serial.print("input value: ");
-        Serial.print(value);
-        Serial.print("\r\n");
-        
-        myservo.write((int)value);
-
-      }
-    }
-  }
-}
-
-
-
 /**
  * Evaluate Serial Input (keys)
+ * TODO: do else-if instead.
  */
-void serialEvaluation() {
-  
-  if(Serial.available()){ //if there's something coming through the serial port
-    val = Serial.read();  //buffer it
+void evaluateSerial() {
+  if (Serial.available() > 0) {
+    char input = Serial.read();
 
-    if(val == servoKey1){
+    if (input == 'q'){
       myservo.write(0);
-      Serial.write("servo at zero!\r\n");
     }
-    else if (val == servoKey2){
+    if (input == 'w') {
+      myservo.write(45);
+    }
+    if (input == 'e') {
       myservo.write(90);
-      Serial.write("servo at center!\r\n");
     }
-    else if (val == servoKey3){
+    if (input == 'r') {
+      myservo.write(135);
+    }
+    if (input == 't') {
       myservo.write(180);
-      Serial.write("servo at max!\r\n");
-    }
-    else if (val == servoKey4){
-      myservo.write(20);
-      Serial.write("servo at arbitrary position. \r\n");  
-    }
-    else {
-      Serial.write("Invalid input!\r\n");
     }
   }
 }
